fix rotation

+++ srcs/raycasting.c
@@ -6,7 +6,7 @@
 /*   By: ericard@student.42.fr <ericard>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/03/04 16:27:13 by ericard@stu       #+#    #+#             */
-/*   Updated: 2021/03/09 15:33:39 by ericard@stu      ###   ########.fr       */
+/*   Updated: 2021/03/11 14:22:17 by ericard@stu      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ void	initial_sidedist(t_infos *infos)
 		else
 		{
 			infos->ray.stepx = 1;
-			infos->ray.sidedistx = (infos->ray.mapx + 1 - infos->ray.posx) * infos->ray.deltadistx;
+			infos->ray.sidedistx = (infos->ray.mapx + 1.0 - infos->ray.posx) * infos->ray.deltadistx;
 		}
 		if (infos->ray.diry < 0)
 		{
@@ -32,7 +32,7 @@ void	initial_sidedist(t_infos *infos)
 		else
 		{
 			infos->ray.stepy = 1;
-			infos->ray.sidedisty = (infos->ray.mapy + 1 - infos->ray.posy) * infos->ray.deltadisty;
+			infos->ray.sidedisty = (infos->ray.mapy + 1.0 - infos->ray.posy) * infos->ray.deltadisty;
 		}
 }
 
@@ -53,14 +53,13 @@ void	distance_wall(t_infos *infos)
 			infos->ray.mapy += infos->ray.stepy;
 			infos->ray.side = 1;
 		}
-
 		if (infos->map[infos->ray.mapy][infos->ray.mapx] == '1')
 			infos->ray.hit = 1;
 	}
 	if (infos->ray.side == 0)
-		infos->ray.perpwalldist = ((double)infos->ray.mapx - infos->ray.posx + (1 - (double)infos->ray.stepx) / 2) / infos->ray.raydirx;
+		infos->ray.perpwalldist = (infos->ray.mapx - infos->ray.posx + (1 - infos->ray.stepx) / 2) / infos->ray.raydirx;
 	else
-		infos->ray.perpwalldist = ((double)infos->ray.mapy - infos->ray.posy + (1 - (double)infos->ray.stepy) / 2) / infos->ray.raydiry;
+		infos->ray.perpwalldist = (infos->ray.mapy - infos->ray.posy + (1 - infos->ray.stepy) / 2) / infos->ray.raydiry;
 }
 
 void	draw_start_end(t_infos *infos)
